Create STM32 BL-RST.png
Random Idea for auto bootload/reset, stm32.

on power up, jump to flash, if in flash after 400ms and mcu gets a reset, jump to bootloader.
Removing and applying power should let the stm jump to user flash (plug and unplug).

so the idea is, the reset button functions as a BL jump after 400ms of applied power. But a full power cycle will always let the stm jump to flash.

c1 will charge slow but discharge fast. unity gain buffer will "buffer" the cap voltage? and the Schmitt trigger wont trigger untill the cap hits 4v.

and no, the stm32 is not a 5v animal, this is a concept schematic.